Make remedies fully pluggable
Previously, the LibXML driver had a large multi-way condition to decide
which remedy (if any) could be applied to some error.  Instead, each remedy
now has a class method which returns true iff it accepts the current driver,
and the constructor is permitted to return nothing if the error does not
look like one the remedy can fix.