New UnclosedHTML remedy
This helps for feeds with a a <description> containing a simplistic substring
of a piece of HTML markup.  When the parser finds a close-tag mismatch where
the unclosed element's name is that of a non-empty HTML element, inject an
extra close-tag at that point.

This change adds a dependency on HTML::Tagset.